refactor(components.table-sort): move to telecomUniverseComponents
