+++ client/components/notification/notification-list.html
@@ -14,21 +14,21 @@
     </button>
     <table class="table table-hover">
         <thead>
-            <tr class="sort" data-table-sort="NotificationListCtrl.sortElements(SORT)">
+            <tr class="sort" data-tuc-table-sort="NotificationListCtrl.sortElements(SORT)">
                 <th scope="col">
-                    <col-sort data-field-name="type" data-title="{{:: 'table_sort' | translate}}">
+                    <tuc-col-sort data-field-name="type" data-title="{{:: 'table_sort' | translate}}">
                         {{:: 'components_notification_type' | translate }}
-                    </col-sort>
+                    </tuc-col-sort>
                 </th>
                 <th scope="col">
-                    <col-sort data-field-name="contact" data-title="{{:: 'table_sort' | translate}}">
+                    <tuc-col-sort data-field-name="contact" data-title="{{:: 'table_sort' | translate}}">
                         {{:: 'components_notification_contact' | translate }}
-                    </col-sort>
+                    </tuc-col-sort>
                 </th>
                 <th scope="col">
-                    <col-sort data-field-name="frequency" data-title="{{:: 'table_sort' | translate}}">
+                    <tuc-col-sort data-field-name="frequency" data-title="{{:: 'table_sort' | translate}}">
                         {{:: 'components_notification_frequency' | translate }}
-                    </col-sort>
+                    </tuc-col-sort>
                 </th>
                 <th scope="col"></th>
             </tr>
