+++ client/telecomUniverseComponents/index.js
@@ -1,5 +1,9 @@
 import angular from 'angular';
 
+import tucTableSort from './table-sort';
+
 export default angular
-  .module('telecomUniverseComponents', [])
+  .module('telecomUniverseComponents', [
+    tucTableSort,
+  ])
   .name;
