+++ client/telecomUniverseComponents/table-sort/table-sort.directive.js
@@ -1,0 +1,15 @@
+import _ from 'lodash';
+
+export default () => ({
+  restrict: 'A',
+  scope: {
+    tucTableSort: '&',
+  },
+  link(scope) {
+    _.set(scope, '$parent.sort', {
+      fieldName: null,
+      descending: null,
+    });
+    _.set(scope, '$parent.tucTableSort', scope.tucTableSort);
+  },
+});
